deriveCollabGate + deriveEditPraxisPhase become leaf modules (#1397)
