+++ frontend/src/pages/praxisDetail/__tests__/unaffiliatedDetailV2.test.tsx
@@ -370,11 +370,17 @@ describe("Unaffiliated praxis detail — the state axes", () => {
   });
 
   it("shows owner controls to a member and nothing to a visitor", () => {
-    expect(render(state()).html, "a visitor gets no edit link").not.toContain(
-      'href="/praxis/1/edit"',
+    // #1397: the cluster is anchored on the UNSUBMIT control now. On a
+    // submitted solo `/edit` redirects straight back to this page, so the edit
+    // link is hidden and unsubmitting is the way into the composer.
+    expect(render(state()).text, "a visitor gets no owner controls").not.toContain(
+      "unsubmit",
     );
     const owner = state({ isOwner: true, user: VIEWER });
-    expect(render(owner).html).toContain('href="/praxis/1/edit"');
+    expect(render(owner).text).toContain("unsubmit");
+    expect(render(owner).html, "and nothing that round-trips").not.toContain(
+      'href="/praxis/1/edit"',
+    );
   });
 
   it("shows the steward bar only in admin mode", () => {
